EVENTLOG: Rewrite cdb2_sqlreplay in C++
